fixed weird delete bug, made patch cords beziers, changed node cursor
The cursor for node and nodetext are now pointer to indicate they are selectable.  The linecap is also now round, which makes thick patch cords look less bad when they join a rect.

Also changed the getXWithId and deleteXWithId functions to check if x[i] is defined before looking at x[i].id.  Turns out WebPd is using id's as indices into the objects array internally, so we need to allow the array to keep growing (ugh).  Maybe better long-term solution is maintain an object that maps indices to ids or vice-versa?